requirements.txt
yt-dlp
python-telegram-bot==20.*
PyYAML
tenacity
requests
pytest-asyncio
